requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
django-cors-headers>=4.0
